main: Clean up PCI_GET_*_ID macros
Properly type-casting their results to 16-bit improves readability. No
functional change, all consumers already did this type-cast implicitly.

+++ main.c
@@ -27,8 +27,8 @@ extern const unsigned long init_array_start[];
 extern const unsigned long init_array_end[];
 extern CHAR16 *boot_medium_path;
 
-#define PCI_GET_VENDOR_ID(id) id
-#define PCI_GET_PRODUCT_ID(id) id >> 16
+#define PCI_GET_VENDOR_ID(id)	(UINT16)(id)
+#define PCI_GET_PRODUCT_ID(id)	(UINT16)((id) >> 16)
 
 static EFI_STATUS probe_watchdogs(EFI_LOADED_IMAGE *loaded_image, UINTN timeout)
 {
